qr/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Pillow>=10.0.0
pyzbar>=0.1.9
python-dotenv>=1.0.0
requests>=2.31.0
Werkzeug>=3.0.0
